+++ Example/LocalizationDemo/SecondViewController.h
@@ -1,0 +1,13 @@
+//
+//  SecondViewController.h
+//  LocalizationDemo
+//
+//  Created by Dmitriy Petrusevich on 22/06/15.
+//  Copyright (c) 2015 Dmitriy Petrusevich. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface SecondViewController : UIViewController
+
+@end